Code/src/requirements.txt
openai>=1.0.0
anthropic>=0.20.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
datasets>=2.14.0
transformers>=4.30.0
torch>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
requests>=2.31.0
krippendorff>=0.6.0
statsmodels>=0.14.0
